backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
python-dotenv
google-generativeai
